requirements.txt
opencv-python
opencv-contrib-python
numpy
Pillow
sqlalchemy
streamlit
tk-tools
pyserial
paho-mqtt==1.5.1
RPi.GPIO
picamera2
pyaudio